feat: исправления в работе программы

+++ src/main/java/com/test/t1/dto/TaskRequest.java
@@ -3,6 +3,6 @@ package com.test.t1.dto;
 import com.test.t1.model.Status;
 import jakarta.annotation.Nullable;
 
-public record TaskRequest(String title, String description, int userId, @Nullable Status status) {
+public record TaskRequest(String title, String description, long userId, @Nullable Status status) {
 
 }
